fix(fluxing): ship v0.18.34.0 - 3 user-reported bugfixes + installer
3 user-reported bugs fixed in 0.18.30.0+:

1. Cannot input Chinese (TSF profile not registered):
   install.nsi:437 now runs regsvr32 weasel.dll (32-bit TSF shim)
   in addition to regsvr32 weaselx64.dll (64-bit TSF shim). Without
   32-bit shim registered, HKCU\Software\Microsoft\CTF\Assemblies\
   0x00000804 had no entry for Fluxing.

2. QuickPanel always shown:
   Removed EnableAlwaysShowMode() from WeaselServerApp.cpp:40.
   spec 052 was implemented as show-on-service-start instead of
   show-on-IME-focus, making the panel appear even when user
   was not using Fluxing IME.

3. QuickPanel UI ugly dark border:
   QuickPanelDialog.cpp DoPaint rewritten: bar border alpha=0
   (transparent), separator lines alpha=0x10 (was 0x14),
   background alpha=0xE8 (was 0xF0), added 1px hairline shadow
   below bar.

Also fixes QuickPanelDialog.cpp L46 GDI+ DrawLine overload
ambiguity (INT cast on float parameter for the (Pen,INT,INT,INT,INT)
overload).

Verified L46 3-path gate ALL PASS:
- xmake -a x86 -m release: 0 errors, 4 binaries built (x86)
- msbuild weasel.sln: 0 errors, 4 pre-existing warnings, 2m22s
- scripts\test-infra\run-test-suite.bat: ALL TESTS PASSED
- L42 byte-verify: 0x001E1E1E in weasel.dll
- L47 byte-verify: QuickPanelDialog.cpp + WeaselServerApp.cpp
  100% CRLF, no 0xC0/0xC1, no BOM
- L09 install.nsi: BOM preserved, 100% CRLF, no 0xC0/0xC1
- L14 arch-verify: WeaselServer.exe + WeaselDeployer.exe +
  WeaselSetup.exe + weasel.dll + rime.dll all Intel i386

Installer: release/fluxing-0.18.34.0-installer.exe
  (43,120,586 bytes,
   SHA256 3552f12c8e83daa142ecdf7c0d2595b6c5fd3efb86d6f10b47d786dec2f0ebd0)

User action: install over 0.18.30.0+ via silent or GUI. Clean
HKLM\SOFTWARE\WOW6432Node\Fluxing InstallDir pollution first
per L54 if upgrading from an older install.

spec 055/TaskTracker [P0]

+++ WeaselServer/QuickPanelDialog.cpp
@@ -291,8 +291,8 @@ void DoPaint(HWND hwnd) {
   // Soft hairline shadow below the bar (1px line, very light).
   {
     Gdiplus::Pen hp(Gdiplus::Color(0x18, 0x00, 0x00, 0x00), 1.0f);
-    g.DrawLine(&hp, 14, (float)crc.bottom - 0.5f,
-               crc.right - 14, (float)crc.bottom - 0.5f);
+    g.DrawLine(&hp, (INT)14, (INT)((float)crc.bottom - 0.5f),
+                   (INT)(crc.right - 14), (INT)((float)crc.bottom - 0.5f));
   }
 
   // Logo: gradient rounded rect + logo image
